Break up LedIdentifier and use factory methods.
Hides the pattern lists as file-static, among other advantages.
