+++ plugins/videobasedtracker/IdentifierHelpers.h
@@ -25,15 +25,58 @@
 #ifndef INCLUDED_IdentifierHelpers_h_GUID_B6F81E02_BE7B_4382_12E5_87296135997D
 #define INCLUDED_IdentifierHelpers_h_GUID_B6F81E02_BE7B_4382_12E5_87296135997D
 
-
 // Internal Includes
-// - none
+#include "Types.h"
 
 // Library/third-party includes
-// - none
+#include <boost/assert.hpp>
 
 // Standard includes
-// - none
+#include <algorithm>
+
+namespace osvr {
+namespace vbtracker {
+
+    /// @brief Helper for implementations of LedIdentifier to truncate the
+    /// passed-in brightness list to the maximum useful length.
+    inline void truncateBrightnessListTo(BrightnessList &brightnesses,
+                                         size_t n) {
+        /// @todo has to be a more efficient method.
+        while (brightnesses.size() > n) {
+            brightnesses.pop_front();
+        }
+    }
+
+    /// @brief Helper function for implementations of LedIdentifier to find
+    /// the minimum and maximum values in a non-empty list of brightnesses.
+    inline BrightnessMinMax
+    findMinMaxBrightness(const BrightnessList &brightnesses) {
+
+        BOOST_ASSERT_MSG(!brightnesses.empty(), "Must be a non-empty list!");
+        auto extremaIterators =
+            std::minmax_element(begin(brightnesses), end(brightnesses));
+        return std::make_pair(*extremaIterators.first,
+                              *extremaIterators.second);
+    }
+
+    /// @brief Helper for implementations of LedIdentifier to turn a
+    /// brightness list into a boolean list based on thresholding on the
+    /// halfway point between minimum and maximum brightness.
+    inline LedPattern getBitsUsingThreshold(const BrightnessList &brightnesses,
+                                            float threshold) {
+        LedPattern ret;
+        // Allocate output space for our transform.
+        ret.resize(brightnesses.size());
+
+        // Transform the brightnesses into a container of bools using this
+        // little lambda
+        std::transform(
+            begin(brightnesses), end(brightnesses), begin(ret),
+            [threshold](Brightness val) { return val >= threshold; });
+
+        return ret;
+    }
+} // End namespace vbtracker
+} // End namespace osvr
 
 #endif // INCLUDED_IdentifierHelpers_h_GUID_B6F81E02_BE7B_4382_12E5_87296135997D
-
